refactor(widget-plugin-rollup-build-ts): use separate file for rollup config
